Code concern separation.
