+++ aquarius/graphql.py
@@ -1,0 +1,57 @@
+import time
+
+from gql import gql, Client
+from gql.transport.aiohttp import AIOHTTPTransport
+
+from aquarius.events.util import get_network_name
+
+
+def get_number_orders(token_address, last_sync_block):
+    client = get_client()
+
+    last_block = get_last_block(client)
+    while last_block <= last_sync_block:
+        last_block = get_last_block(client)
+        time.sleep(2)
+
+    did_query = gql(
+        '{ datatokens(where: {id: "'
+        + token_address.lower()
+        + '"}) { orderVolume } }'
+    )
+    result = client.execute(did_query)
+
+    try:
+        number_orders = result["datatokens"][0]["orderVolume"]
+    except (KeyError, IndexError):
+        raise Exception(
+            "Can not get number of orders for subgraph {get_network_name()} did {did}"
+        )
+
+    return number_orders
+
+
+def get_transport():
+    network_name = get_network_name()
+    if network_name == "ganache":
+        prefix = "http://localhost:9000"
+    else:
+        prefix = f"http://subgraph.{network_name}.oceanprotocol.com"
+
+    return AIOHTTPTransport(url=f"{prefix}/subgraphs/name/oceanprotocol/ocean-subgraph")
+
+
+def get_client():
+    return Client(transport=get_transport(), fetch_schema_from_transport=True)
+
+
+def get_last_block(client):
+    last_block_query = gql("{_meta { block { number } } }")
+
+    try:
+        result = client.execute(last_block_query)
+        last_block = result["_meta"]["block"]["number"]
+    except (KeyError, IndexError):
+        raise Exception("Can not get last block name for subgraph {get_network_name()}")
+
+    return last_block
